Start Implementing support for nested conditionals
