+++ tests/test_conditionals.py
@@ -1,5 +1,6 @@
 import pytest
 from gotranx.expressions import build_expression
+from gotranx.codegen import BaseGotranODECodePrinter
 import sympy as sp
 
 
@@ -27,3 +28,27 @@ def test_Conditional_expr(expr, expected, parser, trans):
     assert sympy_expr.subs({"time": 0}) == expected[0]
     assert sympy_expr.subs({"time": 1}) == expected[1]
     assert sympy_expr.subs({"time": -1}) == expected[2]
+
+
+def test_single_Conditional_from_sympy():
+    t = sp.Symbol("t")
+    a = sp.Symbol("a")
+    b = sp.Symbol("b")
+
+    expr = sp.Piecewise((a, sp.Lt(t, 0)), (b, True))
+    printer = BaseGotranODECodePrinter()
+    result = printer.doprint(expr)
+    assert result == "Conditional(Lt(t, 0), a, b)"
+
+
+def test_nested_Conditional_from_sympy():
+    t = sp.Symbol("t")
+    a = sp.Symbol("a")
+    b = sp.Symbol("b")
+    c = sp.Symbol("c")
+
+    expr = sp.Piecewise((a, t < 0), (b, t > 0), (c, True))
+    printer = BaseGotranODECodePrinter()
+    result = printer.doprint(expr)
+
+    assert result == "Conditional(Lt(t, 0), a, Conditional(Gt(t, 0), b, c))"
